Rendering a Flat Texture in 3D Space.
Changed and moved the 2D texture renderer. It currently is a function
within the internals of the 3D renderer that renderers a texture
(GLuint) within the 3D space. This will be used for billboards in the
future. An actual 2D renderer has not been created. For the 2D renderer,
we will probably use OpenGL as well (or SDL, but that uses OpenGL
anyways).