Finally restructured "controls.hpp" and "controls.cpp".
The controls system was completely restructured. Now there are three new
classes: AelaTimeManager (a simple header-only class for managing the
time), AelaControlManager (a class for managing controls that will be
greatly expanded in the future, especially when key bindings become a
thing) and Aela3DCamera (a class for a camera. Before the camera was
handled using procedural programming in controls.cpp).

*Note*: In some functions I had to pass matrices (glm::mat4) by their
values rather than by their pointers because using pointers caused MAJOR
errors.